MDEV-16207: Sig 11 on RHEL7 on Azure
The problem was that Item_subselect::const_item was not checked correctly and can differ from used_tables result.
Which lead to the situation when Item_direct_view_ref pointing on subselect was constant but the subquery return
FALSE on const_item() request (which was an error).

Now both methods report the same.