MDEV-16188: Fixed several issues/problems.
1. Fixed the bug leading to improper calculation of positions
for key entries in leaf nodes of MyISAM B-trees. Fixed the bug that
that caused in some cases incorrect estimation of the number
records in range due to premature rounding.
2. Changed the implementation of handler::multi_range_read_info_const()
to take into account the width of the gaps between ranges that made
the cost of range scan, especially index only range scans, more accurate.
This is extremely important for proper estimation of the cost of usage
range filters because to build a range filter we use index only range
scan.
3. Changed the class Cost_estimate adding the members that allowed
to calculate separately the cost of accessing index nodes in B-trees.
This was done to be able to figure out whether a scan was good
to be used for a range filter. The cost of index only range scan for
the i-th index is saved now in TABLE::quick_index_only_costs[i].
4. Restricted the usage of the virtual handler function keyread_time().
Added an implementation of this function for ha_partition.
5. Added a virtual handler function key_scan_time() with a default
implementation and a separate implementation for ha_partition.

All these changes required adjustment of many tests and/or test results.